+++ backend/canisters/local_user_index/impl/src/jobs/upgrade_communities.rs
@@ -1,4 +1,4 @@
-use crate::jobs::clear_chunk_store_if_no_pending_upgrades;
+use crate::jobs::refresh_chunk_store;
 use crate::{RuntimeState, mutate_state};
 use constants::min_cycles_balance;
 use ic_cdk_management_canister::CanisterInstallMode;
@@ -7,7 +7,7 @@ use per_round_timer::PerRoundTimer;
 use std::cell::RefCell;
 use tracing::trace;
 use types::{BuildVersion, CanisterId, CommunityId, Cycles, CyclesTopUp};
-use utils::canister::{CanisterToInstall, ChunkedWasmToInstall, FailedUpgrade, WasmToInstall, install};
+use utils::canister::{CanisterToInstall, FailedUpgrade, install};
 
 thread_local! {
     static TIMER: RefCell<Option<PerRoundTimer>> = RefCell::default();
@@ -36,7 +36,7 @@ fn run() {
         TIMER.set(None);
         trace!("'upgrade_communities' job stopped");
 
-        clear_chunk_store_if_no_pending_upgrades();
+        refresh_chunk_store::refresh_if_no_pending_upgrades();
     }
 }
 
@@ -90,15 +90,7 @@ fn initialize_upgrade(canister_id: CanisterId, force: bool, state: &mut RuntimeS
         canister_id,
         current_wasm_version,
         new_wasm_version,
-        new_wasm: if community_canister_wasm.chunks.is_empty() {
-            WasmToInstall::Default(community_canister_wasm.wasm.module.clone())
-        } else {
-            WasmToInstall::Chunked(ChunkedWasmToInstall {
-                chunks: community_canister_wasm.chunks.clone(),
-                wasm_hash: community_canister_wasm.wasm_hash,
-                store_canister_id: state.env.canister_id(),
-            })
-        },
+        new_wasm: state.child_canister_wasm_to_install(ChildCanisterType::Community).wasm,
         deposit_cycles_if_needed,
         args: msgpack::serialize_then_unwrap(&community_canister::post_upgrade::Args {
             wasm_version: new_wasm_version,